+++ feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/ui/modal/NetworkErrorModal.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.ai_chat.ui.component
+package com.emotionstorage.ai_chat.ui.modal
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.Composable
